Entities with .ase have correct draw order + no memory leak when reloading assets
We store pointer to pointer Asset_Ase, so that when we reload the texture, if it's somewhere different on the heap we don't have to worry anymore.

+++ src/utils/extramath.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-inline int pyth_s(float x, float y, float x2, float y2) {
+inline float pyth_s(float x, float y, float x2, float y2) {
     return (x-x2)*(x-x2) + (y-y2)*(y-y2);
 }
